Move web ui link to a new details item
