+++ packages/forklift-console-plugin/locales/en/plugin__forklift-console-plugin.json
@@ -151,6 +151,7 @@
   "Error: this field must be set to a boolean value.": "Error: this field must be set to a boolean value.",
   "ESXi": "ESXi",
   "Expiration date": "Expiration date",
+  "External management system": "External management system",
   "Failed": "Failed",
   "False": "False",
   "Features": "Features",
