+++ packages/forklift-console-plugin/src/modules/Providers/views/details/components/DetailsSection/OvirtDetailsSection.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useForkliftTranslation } from 'src/utils/i18n';
 
-import { V1beta1Provider } from '@kubev2v/types';
 import { DescriptionList } from '@patternfly/react-core';
 
 import { DetailsItem } from '../../../../utils';
@@ -9,33 +8,21 @@ import { DetailsItem } from '../../../../utils';
 import {
   CreatedAtDetailsItem,
   CredentialsDetailsItem,
-  NameAndUiLinkDetailsItem,
+  ExternalManagementLinkDetailsItem,
+  NameDetailsItem,
   NamespaceDetailsItem,
   OwnerDetailsItem,
   TypeDetailsItem,
   URLDetailsItem,
 } from './components';
 import { DetailsSectionProps } from './DetailsSection';
+import { getOvirtProviderWebUILink } from './utils';
 
 export const OvirtDetailsSection: React.FC<DetailsSectionProps> = ({ data }) => {
   const { t } = useForkliftTranslation();
 
   const { provider, permissions } = data;
-
-  /**
-   * A function for auto calculating the RHV UI link.
-   * It extracts the provider's RHV UI link from the RHV URL by searching for the URL's path of
-   * '/ovirt-engine/api[/]' and cutting out the /api[/] path section.
-   * If RHV URL is invalid then an empty string is returned
-   */
-  const getProviderUiContent = (provider: V1beta1Provider): string => {
-    const uiLinkRegexp = new RegExp('(?<=ovirt-engine)\\/api(\\/)*$', 'g');
-    const regexpResult = uiLinkRegexp.exec(provider?.spec?.url);
-
-    return provider?.spec?.url && regexpResult
-      ? provider?.spec?.url.slice(0, uiLinkRegexp.lastIndex - regexpResult[0].length)
-      : '';
-  };
+  const webUILink = getOvirtProviderWebUILink(provider);
 
   return (
     <DescriptionList
@@ -45,14 +32,14 @@ export const OvirtDetailsSection: React.FC<DetailsSectionProps> = ({ data }) => 
     >
       <TypeDetailsItem resource={provider} />
 
-      <DetailsItem title={''} content={''} />
-
-      <NameAndUiLinkDetailsItem
+      <ExternalManagementLinkDetailsItem
         resource={provider}
         canPatch={permissions.canPatch}
         webUILinkText={t(`Red Hat Virtualization Manager UI`)}
-        webUILinkCalcVal={getProviderUiContent(provider)}
+        webUILink={webUILink}
       />
+
+      <NameDetailsItem resource={provider} />
 
       <NamespaceDetailsItem resource={provider} />
