+++ packages/forklift-console-plugin/src/modules/Providers/views/details/components/DetailsSection/index.ts
@@ -1,8 +1,10 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
+export * from './components';
 export * from './DetailsSection';
 export * from './OpenshiftDetailsSection';
 export * from './OpenstackDetailsSection';
 export * from './OVADetailsSection';
 export * from './OvirtDetailsSection';
+export * from './utils';
 export * from './VSphereDetailsSection';
 // @endindex
