+++ packages/forklift-console-plugin/src/modules/Providers/views/details/components/DetailsSection/utils/getProviderUIAnnotation.ts
@@ -1,0 +1,4 @@
+import { V1beta1Provider } from '@kubev2v/types';
+
+export const getProviderUIAnnotation = (provider: V1beta1Provider): string =>
+  provider?.metadata?.annotations?.['forklift.konveyor.io/providerUI'];
