+++ packages/forklift-console-plugin/src/modules/Providers/views/details/components/DetailsSection/utils/index.ts
@@ -1,0 +1,4 @@
+// @index(['./*', /style/g], f => `export * from '${f.path}';`)
+export * from './getOvirtProviderWebUILink';
+export * from './getProviderUIAnnotation';
+// @endindex
